% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLengthHist.R
\name{make_length_plot}
\alias{make_length_plot}
\title{Read length histogram (log10 x-axis)}
\usage{
make_length_plot(readLengths, file_name)
}
\arguments{
\item{readLengths}{Numeric vector (or list coercible to numeric) of read lengths.}
}
\value{
A ggplot object.
}
\description{
Produces a histogram of read lengths with a log10 x-axis.
}
\keyword{internal}
