% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualMat.R
\name{QualMat}
\alias{QualMat}
\title{calculates and fills metrics of LongReadQc object
metrics:
length
distribution
read lengths
yield
GC content
per read
per position          TO DO
per file
N content
per position          TO DO? do we even care
N50
N90
Q score
per read average
per position average <- takes a lot of memory, think of different approach
per file average

ADD ERROR HANDLING
USE SYSTIME TO CHECK HOW LONG IT TAKES
FIGURE OUT WHAT IS TAKING UP SO MUCH MEMORY IN THE OBJECT
save only N count if >0}
\usage{
QualMat(qc_obj, stringset, filename)
}
\description{
use profvis on function to figure out data use stuff
use microbenchmark to figure out how different approaches compare
data table instead of dataframe?
try working with asciis instead of numbres?
try preallocating memory for memory
way to work not with metrics but only
at x position -> take all reads with length greater than x -> calculate quartiles -> next pos
add adapter content as metric
Update workflow
-> try to use workflow function
-> generates what needs to be generated
-> for next week having working workflow
run profvis on R terminal
}
