% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongReadQC-class.R
\name{show,LongReadQC-method}
\alias{show,LongReadQC-method}
\title{Show method for LongReadQC objects}
\usage{
\S4method{show}{LongReadQC}(object)
}
\arguments{
\item{object}{A \code{LongReadQC} object.}
}
\value{
Invisibly returns the input \code{LongReadQC} object.
}
\description{
Displays a concise summary of a \code{LongReadQC} object, including the
number of processed files and whether metrics and plots are available.
}
