% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongReadQC-class.R
\name{show,LongReadQC-method}
\alias{show,LongReadQC-method}
\title{Show method for LongReadQC objects}
\usage{
\S4method{show}{LongReadQC}(object)
}
\arguments{
\item{object}{A LongReadQC object}
}
\description{
Show method for LongReadQC objects
}
